refactor models into their own file
