+++ backend/services/student_school/student_school_model.py
@@ -1,18 +1,4 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_cors import CORS
-from os import environ
-
-# flask configs
-app = Flask(__name__)
-db = SQLAlchemy(app)
-
-CORS(app)
-
-app.config['SQLALCHEMY_DATABASE_URI'] = environ.get('SQLALCHEMY_DATABASE_URI') or 'mysql+mysqlconnector://root@localhost:3308/student_school'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.config['SQLALCHEMY_ENGINE_OPTIONS'] = {'pool_recycle': 299}
-
+from student_school_api import db
 
 class School(db.Model):
     __tablename__ = 'school'
@@ -28,6 +14,7 @@ class School(db.Model):
             "school_id": self.school_id,
             "school_name": self.school_name,
         }
+
 
 class Student(db.Model):
     __tablename__ = 'student'
@@ -48,11 +35,3 @@ class Student(db.Model):
             "student_name": self.student_name,
             "school_id": self.school_id
         }
-
-@app.route('/')
-def hello():
-    return 'Hello, World!'
-
-if __name__ == "__main__":
-    # should be binding to 0.0.0.0 if you want the container to be accessible from outside
-    app.run(host="0.0.0.0", port=6000, debug=True)
